requirements.txt
python-consul
docker